Code formatting; moved all tables to it's own folder
