+++ resources/js/Pages/Constructors/Index.vue
@@ -1,5 +1,6 @@
 <template>
   <Head :title="franchise.name + ' Constructors'" />
+
   <PageHeader :title="franchise.name"/>
 
   <TabBar :tabs="tabs" />
@@ -18,8 +19,20 @@
   })
 
   const tabs = [
-    { name: 'Drivers', href: '/franchise/' + props.franchise.slug + '/drivers', current: false },
-    { name: 'Constructors', href: '/franchise/' + props.franchise.slug + '/constructors', current: true },
-    { name: 'Events', href: '/franchise/' + props.franchise.slug + '/events', current: false },
+    {
+      name: 'Drivers',
+      href: '/franchise/' + props.franchise.slug + '/drivers',
+      current: false
+    },
+    {
+      name: 'Constructors',
+      href: '/franchise/' + props.franchise.slug + '/constructors',
+      current: true
+    },
+    {
+      name: 'Events',
+      href: '/franchise/' + props.franchise.slug + '/events',
+      current: false
+    },
   ]
 </script>
