+++ resources/js/Pages/Constructors/Show.vue
@@ -1,16 +1,17 @@
 <template>
 	<Head :title="team.name" />
+
 	<PageHeader :title="team.name"/>
 
   <Drivers :drivers="drivers" :franchiseSlug="franchise.slug"/>
 
-  <ConstructorsResultsTable title="Results" :franchise="franchise" :results="results" />
+  <ConstructorsResultsTable :franchise="franchise" :results="results" />
 </template>
 
 <script setup>
   import PageHeader from '@/Components/PageHeadings/PageHeader.vue'
   import Drivers from '@/Components/PageSections/Drivers.vue'
-  import ConstructorsResultsTable from '@/Components/Lists/ConstructorsResultsTable.vue'
+  import ConstructorsResultsTable from '@/Components/Tables/ConstructorsResultsTable.vue'
 
   defineProps({
     team: Object,
