+++ resources/js/Pages/Drivers/Show.vue
@@ -1,5 +1,6 @@
 <template>
   <Head :title="driver.first_name + ' ' + driver.last_name" />
+
   <main>
     <div class="flex flex-wrap bg-white mx-8 py-8 px-8 rounded my-8">
       <img class="max-w-md w-full mx-auto object-center object-cover rounded"
@@ -32,7 +33,7 @@
 </template>
 
 <script setup>
-  import ResultsTable from '@/Components/Lists/ResultsTable.vue'
+  import ResultsTable from '@/Components/Tables/ResultsTable.vue'
 
   defineProps({
     driver: Object,
