+++ resources/js/Pages/Home.vue
@@ -1,85 +1,86 @@
 <template>
-    <Head title="Home" />
-    <div class="relative bg-white overflow-hidden">
-        <div class="max-w-7xl mx-auto">
-            <div class="relative z-10 pb-8 bg-white sm:pb-16 md:pb-20 lg:max-w-2xl lg:w-full lg:pb-28 xl:pb-32">
-                <svg
-                    class="hidden lg:block absolute right-0 inset-y-0 h-full w-48 text-white transform translate-x-1/2"
-                    fill="currentColor"
-                    viewBox="0 0 100 100"
-                    preserveAspectRatio="none"
-                    aria-hidden="true"
-                >
-                    <polygon points="50,0 100,0 50,100 0,100" />
-                </svg>
+  <Head title="Home"/>
+  <div class="relative bg-white overflow-hidden">
+    <div class="max-w-7xl mx-auto">
+      <div class="relative z-10 pb-8 bg-white sm:pb-16 md:pb-20 lg:max-w-2xl lg:w-full lg:pb-28 xl:pb-32">
+        <svg
+          class="hidden lg:block absolute right-0 inset-y-0 h-full w-48 text-white transform translate-x-1/2"
+          fill="currentColor"
+          viewBox="0 0 100 100"
+          preserveAspectRatio="none"
+          aria-hidden="true"
+        >
+          <polygon points="50,0 100,0 50,100 0,100"/>
+        </svg>
 
-                <Popover>
-                    <div class="relative pt-6 px-4 sm:px-6 lg:px-8">
-                        <div class="flex items-center flex-grow flex-shrink-0 lg:flex-grow-0">
-                            <div class="flex items-center justify-between w-full md:w-auto">
-                                <div class="-mr-2 flex items-center md:hidden">
-                                    <PopoverButton
-                                        class="bg-white rounded-md p-2 inline-flex items-center justify-center text-gray-400 
+        <Popover>
+          <div class="relative pt-6 px-4 sm:px-6 lg:px-8">
+            <div class="flex items-center flex-grow flex-shrink-0 lg:flex-grow-0">
+              <div class="flex items-center justify-between w-full md:w-auto">
+                <div class="-mr-2 flex items-center md:hidden">
+                  <PopoverButton
+                    class="bg-white rounded-md p-2 inline-flex items-center justify-center text-gray-400
                                         hover:text-gray-500 hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500"
-                                    />
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </Popover>
+                  />
+                </div>
+              </div>
+            </div>
+          </div>
+        </Popover>
 
-                <div class="mt-10 mx-auto max-w-7xl px-4 sm:mt-12 sm:px-6 md:mt-16 lg:mt-20 lg:px-8 xl:mt-28">
-                    <div class="sm:text-center lg:text-left">
-                        <h1 class="text-4xl tracking-tight font-extrabold text-gray-900 sm:text-5xl md:text-6xl">
+        <div class="mt-10 mx-auto max-w-7xl px-4 sm:mt-12 sm:px-6 md:mt-16 lg:mt-20 lg:px-8 xl:mt-28">
+          <div class="sm:text-center lg:text-left">
+            <h1 class="text-4xl tracking-tight font-extrabold text-gray-900 sm:text-5xl md:text-6xl">
                             <span class="block xl:inline">
                                 Online Fantasy Racing <br>
                             </span>
-                            {{ " " }}
-                            <span class="block text-orange-600 xl:inline">
+              {{ " " }}
+              <span class="block text-orange-600 xl:inline">
                                 your way.
                             </span>
-                        </h1>
-                        <p class="mt-3 text-base text-gray-500 sm:mt-5 sm:text-lg sm:max-w-xl sm:mx-auto md:mt-5 md:text-xl lg:mx-0">
-                            Join today, and race your way.
-                        </p>
-                        <div class="mt-5 sm:mt-8 sm:flex sm:justify-center lg:justify-start">
-                            <div class="rounded-md shadow">
-                                <Link :href="route('register')"
-                                    class="w-full flex items-center justify-center px-8 py-3 
+            </h1>
+            <p
+              class="mt-3 text-base text-gray-500 sm:mt-5 sm:text-lg sm:max-w-xl sm:mx-auto md:mt-5 md:text-xl lg:mx-0">
+              Join today, and race your way.
+            </p>
+            <div class="mt-5 sm:mt-8 sm:flex sm:justify-center lg:justify-start">
+              <div class="rounded-md shadow">
+                <Link :href="route('register')"
+                      class="w-full flex items-center justify-center px-8 py-3
                                     border border-transparent text-base font-medium rounded-md text-white bg-orange-600 hover:bg-gray-800 md:py-4 md:text-lg md:px-10"
-                                >
-                                    Join
-                                </Link>
-                            </div>
-                        </div>
-                    </div>
-                </div>
+                >
+                  Join
+                </Link>
+              </div>
             </div>
+          </div>
         </div>
-        <div class="lg:absolute lg:inset-y-0 lg:right-0 lg:w-1/2">
-            <img
-                class="h-56 w-full object-cover sm:h-72 md:h-96 lg:w-full lg:h-full"
-                src="https://images.unsplash.com/photo-1536751048178-14106afab4f9?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=3381&q=80"
-                alt=""
-            />
-        </div>
+      </div>
     </div>
-    <Features />
+    <div class="lg:absolute lg:inset-y-0 lg:right-0 lg:w-1/2">
+      <img
+        class="h-56 w-full object-cover sm:h-72 md:h-96 lg:w-full lg:h-full"
+        src="https://images.unsplash.com/photo-1536751048178-14106afab4f9?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=3381&q=80"
+        alt=""
+      />
+    </div>
+  </div>
+  <Features/>
 </template>
 
 <script>
-import { Popover, PopoverButton, PopoverPanel } from "@headlessui/vue";
-import Features from "@/Components/PageSections/Features.vue";
+import {Popover, PopoverButton, PopoverPanel} from "@headlessui/vue"
+import Features from "@/Components/PageSections/Features.vue"
 
 export default {
-    components: {
-        Popover,
-        PopoverButton,
-        PopoverPanel,
-        Features,
-    },
-    setup() {
-        return {};
-    },
-};
+  components: {
+    Popover,
+    PopoverButton,
+    PopoverPanel,
+    Features,
+  },
+  setup() {
+    return {}
+  },
+}
 </script>
