+++ vite.config.js
@@ -10,6 +10,7 @@ let host = 'openfantasyracing.test'
 export default defineConfig({
 	plugins: [
 		laravel([
+			'resources/css/app.css',
 			'resources/js/app.js',
 		]),
 		vue({
